Made sure ints were being passed to Qt Widgets/Functions

+++ multitargeting.py
@@ -186,7 +186,7 @@ class Multitargeting(QtWidgets.QMainWindow):
             currentHeight = self.size().height()
 
             #make sure chromosome viewer doesnt get too small
-            self.groupBox_2.setMinimumHeight(0.3 * height)
+            self.groupBox_2.setMinimumHeight(int(0.3 * height))
 
             # window scaling
             scaledWidth = int((width * 1400) / 1920)
@@ -198,7 +198,7 @@ class Multitargeting(QtWidgets.QMainWindow):
                 scaledWidth = currentWidth
 
             #set min width of table
-            self.table.setMinimumWidth(0.5 * scaledWidth)
+            self.table.setMinimumWidth(int(0.5 * scaledWidth))
 
             screen = QtWidgets.QApplication.desktop().screenNumber(QtWidgets.QApplication.desktop().cursor().pos())
             centerPoint = QtWidgets.QApplication.desktop().screenGeometry(screen).center()
